MNT: Re-rendered with conda-smithy 2.4.2
Now that this is a noarch Python package, we need to re-render to get
rid of the Python CI matrices. Also there is no need to build this on
multiple CIs. So we build this only on one CI that seems to be the
fastest (CircleCI). Finally we update the README to note this is a
noarch package with all builds done on CircleCI only.